Moved the merge_coverage_data and its validation jobs into new folder struncture. Moved test data/schema. Update Glue script. Updated tests for each job.
