+++ terraform/pipeline/glue.tf
@@ -462,7 +462,7 @@ module "merge_ind_cqc_data_job" {
 
 module "merge_coverage_data_job" {
   source          = "../modules/glue-job"
-  script_dir      = "jobs"
+  script_dir      = "projects/_02_sfc_internal/cqc_coverage/jobs"
   script_name     = "merge_coverage_data.py"
   glue_role       = aws_iam_role.sfc_glue_service_iam_role
   resource_bucket = module.pipeline_resources
@@ -591,7 +591,7 @@ module "validate_merged_ind_cqc_data_job" {
 
 module "validate_merge_coverage_data_job" {
   source          = "../modules/glue-job"
-  script_dir      = "jobs"
+  script_dir      = "projects/_02_sfc_internal/cqc_coverage/jobs"
   script_name     = "validate_merge_coverage_data.py"
   glue_role       = aws_iam_role.sfc_glue_service_iam_role
   resource_bucket = module.pipeline_resources
