+++ tests/test_file_schemas.py
@@ -1383,82 +1383,6 @@ class FilterCleanedValuesSchema:
 
 
 @dataclass
-class MergeCoverageData:
-    clean_cqc_location_for_merge_schema = StructType(
-        [
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.name, StringType(), True),
-            StructField(CQCLClean.postal_code, StringType(), True),
-            StructField(CQCLClean.cqc_sector, StringType(), True),
-            StructField(CQCLClean.care_home, StringType(), True),
-            StructField(CQCLClean.number_of_beds, IntegerType(), True),
-        ]
-    )
-
-    clean_ascwds_workplace_for_merge_schema = StructType(
-        [
-            StructField(AWPClean.ascwds_workplace_import_date, DateType(), True),
-            StructField(AWPClean.location_id, StringType(), True),
-            StructField(AWPClean.master_update_date, DateType(), True),
-            StructField(AWPClean.establishment_id, StringType(), True),
-            StructField(AWPClean.total_staff, IntegerType(), True),
-        ]
-    )
-
-    expected_cqc_and_ascwds_merged_schema = StructType(
-        [
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(AWPClean.ascwds_workplace_import_date, DateType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.name, StringType(), True),
-            StructField(CQCLClean.postal_code, StringType(), True),
-            StructField(CQCLClean.cqc_sector, StringType(), True),
-            StructField(CQCLClean.care_home, StringType(), True),
-            StructField(CQCLClean.number_of_beds, IntegerType(), True),
-            StructField(AWPClean.master_update_date, DateType(), True),
-            StructField(AWPClean.establishment_id, StringType(), True),
-            StructField(AWPClean.total_staff, IntegerType(), True),
-        ]
-    )
-
-    sample_in_ascwds_schema = StructType(
-        [
-            StructField(AWPClean.establishment_id, StringType(), True),
-        ]
-    )
-
-    expected_in_ascwds_schema = StructType(
-        [
-            *sample_in_ascwds_schema,
-            StructField(CoverageColumns.in_ascwds, IntegerType(), True),
-        ]
-    )
-
-    sample_cqc_locations_schema = StructType(
-        [StructField(AWPClean.location_id, StringType(), True)]
-    )
-
-    sample_cqc_ratings_for_merge_schema = StructType(
-        [
-            StructField(AWPClean.location_id, StringType(), True),
-            StructField(CQCRatings.date, StringType(), True),
-            StructField(CQCRatings.overall_rating, StringType(), True),
-            StructField(CQCRatings.latest_rating_flag, IntegerType(), True),
-            StructField(CQCRatings.current_or_historic, StringType(), True),
-        ]
-    )
-
-    expected_cqc_locations_and_latest_cqc_rating_schema = StructType(
-        [
-            *sample_cqc_locations_schema,
-            StructField(CQCRatings.date, StringType(), True),
-            StructField(CQCRatings.overall_rating, StringType(), True),
-        ]
-    )
-
-
-@dataclass
 class LmEngagementUtilsSchemas:
     add_columns_for_locality_manager_dashboard_schema = StructType(
         [
@@ -2959,46 +2883,6 @@ class MLModelMetrics:
     )
 
     r2_metric_schema = predictions_schema
-
-
-@dataclass
-class ValidateMergedCoverageData:
-    cqc_locations_schema = StructType(
-        [
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.name, StringType(), True),
-            StructField(CQCLClean.postal_code, StringType(), True),
-            StructField(CQCLClean.care_home, StringType(), True),
-            StructField(CQCLClean.number_of_beds, IntegerType(), True),
-            StructField(Keys.year, StringType(), True),
-            StructField(Keys.month, StringType(), True),
-            StructField(Keys.day, StringType(), True),
-        ]
-    )
-    merged_coverage_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.ascwds_workplace_import_date, DateType(), True),
-            StructField(IndCQC.name, StringType(), True),
-            StructField(CQCLClean.postal_code, StringType(), True),
-            StructField(IndCQC.care_home, StringType(), True),
-        ]
-    )
-
-    calculate_expected_size_schema = StructType(
-        [
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.name, StringType(), True),
-            StructField(CQCLClean.postal_code, StringType(), True),
-            StructField(CQCLClean.care_home, StringType(), True),
-            StructField(Keys.year, StringType(), True),
-            StructField(Keys.month, StringType(), True),
-            StructField(Keys.day, StringType(), True),
-        ]
-    )
 
 
 @dataclass
